backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
ffmpeg-python
celery[redis]
redis
sqlalchemy
psycopg2-binary
python-dotenv